Free Assets Added
They are placed in the Plugins folder.

Dungeon Floor Traps - Asset from the slab traps mod.
DungeonTools - Asset for generating dungeons. It is used to create a labyrinth.
EnemyRobots - Asset with enemy models.
JUHWAN - Asset with model of Abelisk, which in turn will be used as a teleportation point.
Medieval Models Pack - Asset with items. It is planned to take keys from it.
UnityTechnologies - Asset with a robot that will be played by the player.