Is working on the code.
There is an Updater in the project. The point is that it was meant for the player and enemies, because when mono behemoths are updating themselves, it's harder than updating them from one place.
But there wasn't enough time for enemies.
Most of the code is written on interfaces, because it was expected that some features enemies and player would share together. For example collecting keys or taking damage.
In the end, even the traps were scripted on their own without abstraction.

When starting a scene, Zenject binds the links needed for the game. Such as player, pause, notification.
There is a Labyrinth object on the stage, traps, door, keys are nested in it.